+++ src/main/java/com/aorri2/goodsforyou/user/domain/policy/NewUserPasswordPolicy.java
@@ -1,4 +1,7 @@
-package com.aorri2.goodsforyou.user.domain;
+package com.aorri2.goodsforyou.user.domain.policy;
+
+import com.aorri2.goodsforyou.user.domain.User;
+import com.aorri2.goodsforyou.user.domain.UserPolicy;
 
 public class NewUserPasswordPolicy implements UserPolicy {
 	@Override
