+++ src/main/java/com/aorri2/goodsforyou/user/domain/validator/NewUserValidator.java
@@ -1,7 +1,11 @@
-package com.aorri2.goodsforyou.user.domain;
+package com.aorri2.goodsforyou.user.domain.validator;
 
 import java.util.Collections;
 import java.util.List;
+
+import com.aorri2.goodsforyou.user.domain.User;
+import com.aorri2.goodsforyou.user.domain.UserPolicy;
+import com.aorri2.goodsforyou.user.domain.UserValidator;
 
 public class NewUserValidator implements UserValidator {
 	List<UserPolicy> validityPolicyList = Collections.emptyList();
