+++ src/test/java/com/aorri2/goodsforyou/user/domain/NewUserManagementTest.java
@@ -10,6 +10,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
+import com.aorri2.goodsforyou.user.domain.management.NewUserManagement;
+
 @DisplayName("NewUserManagement 클래스")
 class NewUserManagementTest {
 	/**
